+++ system/expressionengine/libraries/Filemanager.php
@@ -4,10 +4,10 @@
  * ExpressionEngine - by EllisLab
  *
  * @package		ExpressionEngine
- * @author		ExpressionEngine Dev Team
+ * @author		EllisLab Dev Team
  * @copyright	Copyright (c) 2003 - 2012, EllisLab, Inc.
- * @license		http://expressionengine.com/user_guide/license.html
- * @link		http://expressionengine.com
+ * @license		http://ellislab.com/expressionengine/user-guide/license.html
+ * @link		http://ellislab.com
  * @since		Version 2.0
  * @filesource
  */
@@ -20,8 +20,8 @@
  * @package		ExpressionEngine
  * @subpackage	Core
  * @category	Filemanager
- * @author		ExpressionEngine Dev Team
- * @link		http://expressionengine.com
+ * @author		EllisLab Dev Team
+ * @link		http://ellislab.com
  */
 
 class Filemanager {
@@ -757,6 +757,15 @@ class Filemanager {
 	 */
 	function setup()
 	{
+		// Make sure there are directories
+		$dirs = $this->directories(FALSE, TRUE);
+		if (empty($dirs))
+		{
+			return $this->EE->output->send_ajax_response(array(
+				'error' => lang('no_upload_dirs')
+			));
+		}
+
 		if (REQ != 'CP')
 		{
 			$this->EE->load->helper('form');
@@ -810,7 +819,7 @@ class Filemanager {
 	public function datatables($first_dir = NULL)
 	{
 		$this->EE->load->model('file_model');
-		
+
 		// Argh
 		$this->EE->_mcp_reference = $this;
 		
@@ -880,12 +889,17 @@ class Filemanager {
 			$file_params['search_value']	= $params['keywords'];
 			$file_params['search_in']		= 'all';
 		}
-		
+
+		// Mask the URL if we're coming from the CP
+		$sync_files_url = (REQ == "CP") ?
+			$this->EE->cp->masked_url('http://ellislab.com/expressionengine/user-guide/cp/content/files/sync_files.html') :
+			'http://ellislab.com/expressionengine/user-guide/cp/content/files/sync_files.html';
+
 		return array(
 			'rows'			=> $this->_browser_get_files($dir, $file_params),
-			'no_results' 	=> sprintf(
+			'no_results'	=> sprintf(
 				lang('no_uploaded_files'), 
-				$this->EE->cp->masked_url('http://expressionengine.com/user_guide/cp/content/files/sync_files.html'),
+				$sync_files_url,
 				BASE.AMP.'C=content_files'.AMP.'M=file_upload_preferences'
 			),
 			'pagination' 	=> array(
@@ -1145,16 +1159,35 @@ class Filemanager {
 		// Channel may not be set for pngs - so we default to highest
 		$image_info['channels'] = ( ! isset($image_info['channels'])) ? 4 : $image_info['channels'];
 
-		$memory_needed = round(($image_info[0] * $image_info[1]
-											  // bits may not always be present
-											* (isset($image_info['bits']) ? $image_info['bits'] : 8)
-											* $image_info['channels'] / 8
-											+ $k64
-								) * $this->_memory_tweak_factor
-                         );
+		$memory_needed = round((
+			$image_info[0] * $image_info[1]
+			// bits may not always be present
+			* (isset($image_info['bits']) ? $image_info['bits'] : 8)
+			* $image_info['channels'] / 8
+			+ $k64
+			) * $this->_memory_tweak_factor
+		);
 
-		$memory_setting = (ini_get('memory_limit') != '') ? intval(ini_get('memory_limit')) : 8;
-		$current = $memory_setting*1024*1024;
+		$memory_setting = ini_get('memory_limit');
+
+		if ( ! $memory_setting)
+		{
+			$memory_setting = '8M';
+		}
+		
+		list($current, $unit) = sscanf($memory_setting, "%d %s");
+
+		switch (strtolower($unit))
+		{
+			case 'g':
+				$current *= 1024;
+				// no break
+			case 'm':
+				$current *= 1024;
+				// no break;
+			case 'k':
+				$current *= 1024;
+		}
 
 		if (function_exists('memory_get_usage'))
 		{
@@ -1346,10 +1379,45 @@ class Filemanager {
 				if ($prefs['width'] > $prefs['height'])
 				{
 					$config['width'] = round($prefs['width'] * $size['height'] / $prefs['height']);
+					
+					// If the new width ends up being smaller than the
+					// resized width
+					if ($config['width'] < $size['width'])
+					{
+						$config['width'] = $size['width'];
+						$config['master_dim'] = 'width';
+					}
 				}
 				elseif ($prefs['height'] > $prefs['width'])
 				{
 					$config['height'] = round($prefs['height'] * $size['width'] / $prefs['width']);
+					
+					// If the new height ends up being smaller than the
+					// desired resized height
+					if ($config['height'] < $size['height'])
+					{
+						$config['height'] = $size['height'];
+						$config['master_dim'] = 'height';
+					}
+				}
+				// If we're dealing with a perfect square image
+				elseif ($prefs['height'] == $prefs['width'])
+				{
+					// And the desired image is landscape, edit the
+					// square image's width to fit
+					if ($size['width'] > $size['height'] ||
+						$size['width'] == $size['height'])
+					{
+						$config['width'] = $size['width'];
+						$config['master_dim'] = 'width';
+					}
+					// If the desired image is portrait, edit the
+					// square image's height to fit
+					elseif ($size['width'] < $size['height'])
+					{
+						$config['height'] = $size['height'];
+						$config['master_dim'] = 'height';
+					}
 				}
 				
 				// First resize down to smallest possible size (greater of height and width)
@@ -1363,8 +1431,8 @@ class Filemanager {
 				// Next set crop accordingly
 				$resized_image_dimensions = $this->get_image_dimensions($resized_path.$prefs['file_name']);
 				$config['source_image'] = $resized_path.$prefs['file_name'];
-				$config['x_axis'] = (($resized_image_dimensions['width'] / 2) - ($config['width'] / 2));
-				$config['y_axis'] = (($resized_image_dimensions['height'] / 2) - ($config['height'] / 2));
+				$config['x_axis'] = (($resized_image_dimensions['width'] / 2) - ($size['width'] / 2));
+				$config['y_axis'] = (($resized_image_dimensions['height'] / 2) - ($size['height'] / 2));
 				$config['maintain_ratio'] = FALSE;
 				
 				// Change height and width back to the desired size
@@ -1380,6 +1448,8 @@ class Filemanager {
 			}
 			else
 			{
+				$config['master_dim'] = $force_master_dim;
+				
 				$this->EE->image_lib->initialize($config);
 				
 				if ( ! $this->EE->image_lib->resize())
@@ -1768,6 +1838,8 @@ class Filemanager {
 
 		foreach ($files as &$file)
 		{
+			$file['file_name'] = urlencode($file['file_name']);
+			
 			// Get thumb information
 			$thumb_info = $this->get_thumb($file, $dir['id']);
 			
@@ -1780,7 +1852,7 @@ class Filemanager {
 					title="'.$file['file_name'].'" 
 					onclick="$.ee_filebrowser.placeImage('.$file['file_id'].'); return false;"
 				>
-					'.$file['file_name'].'
+					'.urldecode($file['file_name']).'
 				</a>';
 			
 			$file['short_name']		= ellipsize($file['title'], 13, 0.5);
@@ -1981,7 +2053,7 @@ class Filemanager {
 			'file_name'		=> $clean_filename,
 			'upload_path'	=> $dir['server_path'],
 			'allowed_types'	=> $allowed_types,
-			'max_size'		=> round($dir['max_size']/1024, 2)
+			'max_size'		=> round($dir['max_size']/1024, 3)
 		);
 		
 		$this->EE->load->helper('xss');
@@ -2270,6 +2342,13 @@ class Filemanager {
 		
 		// Delete the existing file's raw files, but leave the database record
 		$this->EE->file_model->delete_raw_file($file_name, $directory_id);
+
+		// It is possible the file exists but is NOT in the DB yet
+		if (empty($existing_file))
+		{
+			$new_file->modified_by_member_id = $this->EE->session->userdata('member_id');
+			return $new_file;
+		}
 		
 		// Delete the new file's database record, but leave the files
 		$this->EE->file_model->delete_files($new_file->file_id, FALSE);
@@ -2282,6 +2361,7 @@ class Filemanager {
 			'modified_date'			=> $new_file->modified_date,
 			'modified_by_member_id'	=> $this->EE->session->userdata('member_id')
 		));
+		
 		$existing_file->file_size				= $new_file->file_size;
 		$existing_file->file_hw_original		= $new_file->file_hw_original;
 		$existing_file->modified_date			= $new_file->modified_date;
@@ -2318,7 +2398,7 @@ class Filemanager {
 		$config = array(
 			'upload_path'	=> $upload_directory['server_path'],
 			'allowed_types'	=> ($this->EE->session->userdata('group_id') == 1) ? 'all' : $upload_directory['allowed_types'],
-			'max_size'		=> round($upload_directory['max_size']/1024, 2),
+			'max_size'		=> round($upload_directory['max_size']/1024, 3),
 			'max_width'		=> $upload_directory['max_width'],
 			'max_height'	=> $upload_directory['max_height']
 		);
